machine-learning-client/requirements.txt
Flask==2.2.2
numpy==1.20.1
Pillow==9.3.0
pymongo==4.3.3
pytest==6.2.3
python-dotenv==0.21.0
tensorflow==2.11.0
